Allocate the LuaWriter on the stack
Also made a few more methods static and renamed mMapDir to mDir since it
can also be the tileset directory.
